now, disabled timeline is a bit darkened.
git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@1482 3b3fb362-3133-0410-aa15-cf69e0a59cb7

+++ libs/hydrogen/src/hydrogen.cpp
@@ -2115,7 +2115,7 @@ void Hydrogen::addRealtimeNote( int instrument,
 							continue;
 						}
 		
-						if( ( pNote->m_bJustRecorded == false ) && (pNote->get_position() >= postdelete && pNote->get_position() <column + predelete +1 )){
+						if( ( pNote->m_bJustRecorded == false ) && (static_cast<int>( pNote->get_position() ) >= postdelete && pNote->get_position() < column + predelete +1 )){
 							delete pNote;
 							currentPattern->note_map.erase( pos0 );
 						}
@@ -2136,7 +2136,7 @@ void Hydrogen::addRealtimeNote( int instrument,
 					continue;
 				}
 
-				if( ( pNote->m_bJustRecorded == false ) && (pNote->get_position() >= postdelete && pNote->get_position() <column + predelete +1 )){
+				if( ( pNote->m_bJustRecorded == false ) && ( static_cast<int>( pNote->get_position() ) >= postdelete && pNote->get_position() <column + predelete +1 )){
 					delete pNote;
 					currentPattern->note_map.erase( pos0 );
 				}
